requirements.txt
# Core scientific computing
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Structural biology and bioinformatics
biopython>=1.81
biotite>=0.37.0

# MD simulation analysis
MDAnalysis>=2.4.0

# Machine learning for protein engineering
scikit-learn>=1.3.0

# Protein structure analysis
prody>=2.4.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
argparse>=1.4.0

# Jupyter for notebooks
jupyter>=1.0.0
ipykernel>=6.23.0
